Refactor socket endpoint creation into UnixDomainSocket class.
